Code-Clean Up
Made minor tweaks to growth rates.
Cleaned up code pulling classes into their own files, removing some junk, etc.
